pyproject.toml
[project]
name = "aether"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "django-two-factor-auth[phonenumbers]>=1.15.5",
    "django-ratelimit>=4.1.0",
    "Mastodon.py>=1.8.1",
    "atproto>=0.0.55",
    "django-cryptography>=1.1",
]

[dependency-groups]
dev = [
    "djade>=1.5.0",
]
